// apps/web/packages/shared/src/_api/api.ts
export interface GetProfileResponse {
    /**
     * 
     * @type {string}
     * @memberof GetProfileResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof GetProfileResponse
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof GetProfileResponse
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof GetProfileResponse
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof GetProfileResponse
     */
    phone: string | null;
    /**
     * 
     * @type {UserStatus}
     * @memberof GetProfileResponse
     */
    status: UserStatus;
    /**
     * 
     * @type {boolean}
     * @memberof GetProfileResponse
     */
    isGuest: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof GetProfileResponse
     */
    isEmailVerified: boolean;
    /**
     * 
     * @type {Array<GetUserRoleResponse>}
     * @memberof GetProfileResponse
     */
    roles: Array<GetUserRoleResponse>;
}
/**
 * 
 * @export
 * @interface GetUserRoleResponse
 */
export interface GetUserRoleResponse {
    /**
     * 
     * @type {string}
     * @memberof GetUserRoleResponse
     */
    userRoleId: string;
    /**
     * 
     * @type {string}
     * @memberof GetUserRoleResponse
     */
    role: string;
}
// ...
export enum UserStatus {
    Active = 'Active',
    Deactivated = 'Deactivated',
    Inactive = 'Inactive',
    Invited = 'Invited'
}